Add in-browser tests using Karma and Mocha/Chai.
Run tests with:

    npm test

Note: this replaces the previous command which linted and built the
project as a way to test it.

To have tests running in the background, and automatically re-run as
soon as a file changes, use:

    npm run test:watch

By default, Karma will launch Firefox to run tests.
You can have Karma not launch the browser, and load the tests manually,
with:

    npm run test:watch -- --no-browsers

or override the browser(s) that Karma will launch:

    npm run test:watch -- --browsers Chrome,Firefox

(currently, on the Firefox and Chrome launchers are added as dependencies)
Those arguments can also be passed to 'npm test' for single runs.